+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #Changelog
 
 ## upcoming
+- group preview slots now contain media items
 - improve suggested users, get rid of default suggested user, use newest instead
 - add new healthcheck endpoint
 - add endpoint for client version checks